# specs/001-perf-problem-simulator/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Performance Problem Simulator API
  description: |
    API for triggering and monitoring intentional performance problems in an Azure App Service
    application for educational and diagnostic demonstration purposes.
    
    ⚠️ **WARNING**: This API intentionally creates performance problems. It is designed for
    testing and learning purposes only. Do not use in production environments.
    
    ## Endpoint Categories
    
    - **Problem Triggers**: Endpoints that start performance problems
    - **Metrics**: Endpoints that report current application state
    - **Control**: Endpoints for managing simulations
    
    ## API Style
    
    This API uses action-oriented RPC-style endpoints (e.g., `/api/trigger-high-cpu`) rather than
    REST resource-based endpoints, as the operations are actions rather than resource management.
  version: 1.0.0
  contact:
    name: Performance Problem Simulator
  license:
    name: MIT
servers:
  - url: https://localhost:5001
    description: Local development
  - url: https://{appname}.azurewebsites.net
    description: Azure App Service
    variables:
      appname:
        default: perfsimasp48
        description: Azure App Service name
...
paths:
...
  /api/allocate-memory:
    post:
      tags:
        - Memory
      summary: Allocate memory to increase pressure
      description: |
        Allocates the specified amount of memory and holds it until explicitly released.
        Multiple calls are additive - each allocation adds to the total memory held.
        
        **Educational Note**: Memory is allocated as pinned byte arrays on the Large Object Heap
        (LOH), preventing garbage collection until explicitly released.
      operationId: allocateMemory
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoryAllocationRequest'
            examples:
              default:
                summary: Default 100 MB allocation
                value:
                  sizeMegabytes: 100
              large:
                summary: Large 500 MB allocation
                value:
                  sizeMegabytes: 500
      responses:
        '200':
          description: Memory allocated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationResult'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Problem endpoints are disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/release-memory:
    post:
      tags:
        - Memory
      summary: Release all allocated memory
      description: |
        Releases all memory previously allocated via `/api/allocate-memory` and optionally
        forces a garbage collection cycle.
        
        **Educational Note**: After release, memory may not immediately return to baseline
        as the GC runs on its own schedule. Forcing GC can help demonstrate memory recovery.
      operationId: releaseMemory
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                forceGarbageCollection:
                  type: boolean
                  default: true
                  description: Whether to force a Gen2 garbage collection
      responses:
        '200':
          description: Memory released
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationResult'
        '503':
          description: Problem endpoints are disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/trigger-sync-over-async:
    post:
      tags:
        - ThreadPool
      summary: Trigger thread pool starvation via sync-over-async
      description: |
        Demonstrates the sync-over-async anti-pattern by blocking ThreadPool threads.
        This causes thread pool starvation, leading to increased response times for all
        requests as the pool struggles to serve incoming work.
        
        **Educational Note**: This endpoint intentionally uses `.Result` to block threads,
        which is a common anti-pattern. Watch the thread pool metrics to see thread count
        climb while queue depth grows.
      operationId: triggerSyncOverAsync
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThreadBlockRequest'
            examples:
              default:
                summary: Default blocking scenario
                value:
                  delayMilliseconds: 1000
                  concurrentRequests: 10
              aggressive:
                summary: Aggressive starvation
                value:
                  delayMilliseconds: 5000
                  concurrentRequests: 100
      responses:
        '200':
          description: Thread blocking simulation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationResult'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Problem endpoints are disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    MemoryAllocationRequest:
      type: object
      properties:
        sizeMegabytes:
          type: integer
          minimum: 10
          maximum: 1024
          default: 100
          description: Amount of memory to allocate in megabytes
    ThreadBlockRequest:
      type: object
      properties:
        delayMilliseconds:
          type: integer
          minimum: 100
          maximum: 30000
          default: 1000
          description: How long each blocked thread waits
        concurrentRequests:
          type: integer
          minimum: 1
          maximum: 200
          default: 10
          description: Number of concurrent blocking operations
    SimulationResult:
      type: object
      required:
        - simulationId
        - type
        - status
        - message
        - startedAt
      properties:
        simulationId:
          type: string
          format: uuid
          description: Unique identifier for this simulation
        type:
          type: string
          enum: [Cpu, Memory, ThreadBlock]
          description: Type of simulation
        status:
          type: string
          enum: [Started, Completed, Failed]
          description: Current status
        message:
          type: string
          description: Human-readable status message
        actualParameters:
          type: object
          additionalProperties: true
          description: Parameters actually used (may differ from request if values were capped)
        startedAt:
          type: string
          format: date-time
          description: When the simulation started
        estimatedEndAt:
          type: string
          format: date-time
          nullable: true
          description: When the simulation will complete (null for memory allocations)
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Error code
          example: VALIDATION_ERROR
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Field-level validation errors
        timestamp:
          type: string
          format: date-time
...